requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Django==4.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
pyTelegramBotAPI==4.12.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
